Changed Loading screen implementation to check for Movenet every scene
