+++ Unity/Assets/Scenes/HelpMenu.unity
@@ -507,7 +507,7 @@ Transform:
   m_GameObject: {fileID: 994016528}
   m_LocalRotation: {x: 0.40821788, y: -0.23456968, z: 0.10938163, w: 0.8754261}
   m_LocalPosition: {x: 0, y: 3, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_LocalScale: {x: 1, y: 1.0000001, z: 0.99999994}
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 2
@@ -874,6 +874,7 @@ GameObject:
   - component: {fileID: 1441735332}
   - component: {fileID: 1441735336}
   - component: {fileID: 1441735337}
+  - component: {fileID: 1441735338}
   m_Layer: 5
   m_Name: Canvas
   m_TagString: Untagged
@@ -990,6 +991,18 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   GameManagerScript: {fileID: 1238596421}
+--- !u!114 &1441735338
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1441735331}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 5c98f7bdea4cb454fb0a7a88fcb4ac22, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &1449184562
 GameObject:
   m_ObjectHideFlags: 0
