+++ Unity/Assets/Scenes/LevelSelector.unity
@@ -2301,6 +2301,7 @@ GameObject:
   - component: {fileID: 1441735332}
   - component: {fileID: 1441735337}
   - component: {fileID: 1441735336}
+  - component: {fileID: 1441735338}
   m_Layer: 5
   m_Name: Canvas
   m_TagString: Untagged
@@ -2417,6 +2418,18 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   GameManagerScript: {fileID: 1238596421}
+--- !u!114 &1441735338
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1441735331}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 5c98f7bdea4cb454fb0a7a88fcb4ac22, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &1448099759
 GameObject:
   m_ObjectHideFlags: 0
