+++ Unity/Assets/Scripts/mqtt_MovenetConnected.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5c98f7bdea4cb454fb0a7a88fcb4ac22
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
